fix dependency cycle WIP
