+++ apps/server/src/modules/tool/context-external-tool/service/context-external-tool.service.ts
@@ -49,10 +49,4 @@ export class ContextExternalToolService {
 
 		return contextExternalTools;
 	}
-
-	public getToolContextTypes(): ToolContextTypesList {
-		const toolContextTypes: ToolContextTypesList = { data: Object.values(ToolContextType) };
-
-		return toolContextTypes;
-	}
 }
