+++ apps/server/src/modules/tool/external-tool/controller/dto/response/index.ts
@@ -7,3 +7,4 @@ export * from './context-external-tool-configuration-template-list.response';
 export * from './school-external-tool-configuration-template.response';
 export * from './school-external-tool-configuration-template-list.response';
 export * from './external-tool-metadata.response';
+export * from './tool-context-types-list';
