+++ apps/server/src/modules/tool/external-tool/controller/tool-configuration.controller.ts
@@ -20,9 +20,8 @@ import {
 	SchoolExternalToolConfigurationTemplateResponse,
 	SchoolExternalToolIdParams,
 	SchoolIdParams,
+	ToolContextTypesList,
 } from './dto';
-import { ToolContextType } from '../../common/enum';
-import { ToolContextTypesList } from './dto/response/tool-context-types-list';
 
 @ApiTags('Tool')
 @Authenticate('jwt')
