+++ apps/server/src/modules/tool/external-tool/uc/external-tool-configuration.uc.spec.ts
@@ -672,7 +672,7 @@ describe('ExternalToolConfigurationUc', () => {
 
 				await uc.getToolContextTypes(userId);
 
-				expect(contextExternalToolService.getToolContextTypes).toHaveBeenCalled();
+				expect(externalToolConfigurationService.getToolContextTypes).toHaveBeenCalled();
 			});
 		});
 	});
